+++ types/model.ts
@@ -16,12 +16,3 @@ export type ModelProvider = "openai" | "anthropic";
 export type ClientOptions = OpenAIClientOptions | AnthropicClientOptions;
 
 export type ToolCall = AnthropicTool | OpenAITool;
-
-
-export type TextAnnotation = {
-  text: string;
-  midpoint: { x: number; y: number };
-  midpoint_normalized: { x: number; y: number };
-  width: number;
-  height: number;
-}
